scripts/requirements.txt
elevenlabs>=1.0.0
pydub>=0.25.1
pydantic>=2.0.0
python-dotenv>=1.0.0
